% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMIXfunctions.R
\name{binary_group_ind}
\alias{binary_group_ind}
\title{Binary group membership matrix}
\usage{
binary_group_ind(class, G)
}
\arguments{
\item{class}{Numeric vector of class memberships.}

\item{G}{Number of possible different classes.}
}
\value{
Numeric \code{length(class)}\eqn{\times}{x}\eqn{G} matrix of binary group memberships.
}
\description{
Construct the binary group membership matrix from the multinomial classification vector.
}
\examples{

binary_group_ind(class=c(3,1,5), G=6)

}
\author{
Cristina Mollica and Luca Tardella
}
