% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chisqmeasureobs1dim}
\alias{chisqmeasureobs1dim}
\title{Chi-squared index relying on top1 preferences for observed data}
\usage{
chisqmeasureobs1dim(pi_inv, p, weights)
}
\arguments{
\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}

\item{p}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific support parameters.}

\item{weights}{Numeric vector of the \eqn{G} mixture weights.}
}
\value{
Chi-squared index value.
}
\description{
Chi-squared index relying on top1 preferences for observed data
}
