% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chisqmeasureobsmatrix1dim}
\alias{chisqmeasureobsmatrix1dim}
\title{Generic term of the conditional Chi-squared index relying on top1 preferences for observed data}
\usage{
chisqmeasureobsmatrix1dim(pi_inv, p, weights)
}
\arguments{
\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}

\item{p}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific support parameters.}

\item{weights}{Numeric vector of the \eqn{G} mixture weights.}
}
\value{
Numeric \eqn{K}\eqn{\times}{x}\eqn{K} matrix of the conditional Chi-squared index.
}
\description{
Generic term of the conditional Chi-squared index relying on top1 preferences for observed data
}
