% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chisqmeasuretheo}
\alias{chisqmeasuretheo}
\title{Chi-squared index relying on paired comparisons for replicated data}
\usage{
chisqmeasuretheo(N, ref_order, p, weights, pi_inv_obs)
}
\arguments{
\item{N}{Number of sample units.}

\item{ref_order}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific reference orders.}

\item{p}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific support parameters.}

\item{weights}{Numeric vector of the \eqn{G} mixture weights.}

\item{pi_inv_obs}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings (to replicate the observed missingness patterns).}
}
\value{
Chi-squared index value.
}
\description{
Chi-squared index relying on paired comparisons for replicated data
}
