% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CompProbZpartial}
\alias{CompProbZpartial}
\title{Multinomial probability computation for the Multinomial full-conditionals of the latent component memberships}
\usage{
CompProbZpartial(p, pi_inv, Y, u_bin, n_rank, omega)
}
\arguments{
\item{p}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific support parameters.}

\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}

\item{Y}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} matrix of the quantitative latent variables.}

\item{u_bin}{Binary \eqn{N}\eqn{\times}{x}\eqn{K} matrix indicating whether sample unit \eqn{s} ranked item \eqn{i}.}

\item{n_rank}{Numeric vector of length \eqn{N} with the number of items ranked by each sample unit.}

\item{omega}{Numeric vector of the \eqn{G} mixture weights.}
}
\value{
Numeric \eqn{N}\eqn{\times}{x}\eqn{G} matrix of multinomial probabilities.
}
\description{
The function \code{CompProbZpartial} computes the multinomial probabilities of the Multinomial full-conditionals of the latent component memberships for the Gibbs sampling of a Bayesian mixture of Plackett-Luce models.
}
