% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CompRateP}
\alias{CompRateP}
\title{Rate parameter computation for the Gamma full-conditionals of the support parameters}
\usage{
CompRateP(pi_inv, Y, z, u_bin, n_rank, rate0)
}
\arguments{
\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}

\item{Y}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} matrix of the quantitative latent variables.}

\item{z}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} matrix of binary component memberships.}

\item{u_bin}{Binary \eqn{N}\eqn{\times}{x}\eqn{K} matrix indicating whether sample unit \eqn{s} ranked item \eqn{i}.}

\item{n_rank}{Numeric vector of length \eqn{N} with the number of items ranked by each sample unit.}

\item{rate0}{Numeric vector of \eqn{G} rate hyperparameters.}
}
\value{
Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of rate parameters.
}
\description{
The function \code{CompRateP} computes the rate parameters of the Gamma full-conditionals of the support parameters for the Gibbs sampling of a Bayesian mixture of Plackett-Luce models.
}
