% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CompRateYpartial}
\alias{CompRateYpartial}
\title{Rate parameter computation for the Exponential full-conditionals of the quantitative latent variables}
\usage{
CompRateYpartial(p, pi_inv, ref_order, z, n_rank)
}
\arguments{
\item{p}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific support parameters.}

\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}

\item{ref_order}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific reference orders.}

\item{z}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} matrix of binary component memberships.}

\item{n_rank}{Numeric vector of length \eqn{N} with the number of items ranked by each sample unit.}
}
\value{
Numeric \eqn{N}\eqn{\times}{x}\eqn{K} matrix of rate parameters.
}
\description{
The function \code{CompRateYpartial} computes the rate parameters of the Exponential full-conditionals of the quantitative latent variables for the Gibbs sampling of a Bayesian mixture of Plackett-Luce models.
}
