% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estep}
\alias{Estep}
\title{E-step in the EM algorithm for MAP estimation of a Bayesian mixture of Plackett-Luce models}
\usage{
Estep(p, ref_order, weights, pi_inv)
}
\arguments{
\item{p}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific support parameters.}

\item{ref_order}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific reference orders.}

\item{weights}{Numeric vector of the \eqn{G} mixture weights.}

\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}
}
\value{
Numeric \eqn{N}\eqn{\times}{x}\eqn{G} matrix of estimated posterior component membership probabilities.
}
\description{
The function \code{Estep} updates the posterior component membership probabilities in the EM algorithm for MAP estimation of a Bayesian mixture of Plackett-Luce models.
}
