% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMIXfunctions.R
\name{fill_single_entries}
\alias{fill_single_entries}
\title{Utility to fill in single missing entries of top-(K-1) sequences in partial ordering/ranking datasets}
\usage{
fill_single_entries(data)
}
\arguments{
\item{data}{Numeric data matrix of partial sequences.}
}
\value{
Numeric data matrix of partial sequences in the same format of the input \code{data} with possible single missing entries filled.
}
\description{
Utility to fill in single missing entries of top-(K-1) sequences in partial ordering/ranking datasets
}
\author{
Cristina Mollica and Luca Tardella
}
