% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsPLMIX_with_norm.R
\name{gibbsPLMIX_with_norm}
\alias{gibbsPLMIX_with_norm}
\title{gibbsPLMIX_with_norm}
\usage{
gibbsPLMIX_with_norm(
  pi_inv,
  K,
  G,
  init = list(z = NULL, p = NULL),
  n_iter = 1000,
  n_burn = 500,
  hyper = list(shape0 = matrix(1, nrow = G, ncol = K), rate0 = rep(0.001, G), alpha0 =
    rep(1, G)),
  centered_start = FALSE
)
}
\arguments{
\item{pi_inv}{}

\item{K}{}

\item{G}{}

\item{init}{}

\item{n_iter}{}

\item{n_burn}{}

\item{hyper}{}

\item{centered_start}{}
}
\value{

}
\description{
gibbsPLMIX_with_norm
}
