% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{howmanyranked}
\alias{howmanyranked}
\title{Counts how many items are ranked in a partial ranking matrix}
\usage{
howmanyranked(pi_inv)
}
\arguments{
\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}
}
\value{
Numeric vector of length \eqn{N} with the number of items ranked by each sample unit
}
\description{
Counts how many items are ranked in a partial ranking matrix
}
