% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMIXfunctions.R
\name{myorder}
\alias{myorder}
\title{Utility to switch from a partial ranking to a partial ordering (missing positions denoted with zero)}
\usage{
myorder(x)
}
\arguments{
\item{x}{Numeric integer vector}
}
\description{
Utility to switch from a partial ranking to a partial ordering (missing positions denoted with zero)
}
\author{
Cristina Mollica and Luca Tardella
}
