\name{plmix-package}
\alias{plmix-package}
\alias{plmix}
\docType{package}
\title{
\packageTitle{plmix}
}
\description{
\packageDescription{plmix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plmix}
\packageIndices{plmix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{plmix}

Maintainer: \packageMaintainer{plmix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
