% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMIXfunctions.R
\name{print.summary.gsPLMIX}
\alias{print.summary.gsPLMIX}
\title{Print of the summary of Gibbs sampling simulation of a Bayesian mixture of Plackett-Luce models.}
\usage{
\method{print}{summary.gsPLMIX}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary.gsPLMIX} returned by the \code{summary.gsPLMIX} function.}

\item{...}{Further arguments passed to or from other methods (not used).}
}
\description{
\code{print} method for class \code{summary.gsPLMIX}. It shows some general information on the Gibbs sampling simulation of a Bayesian mixture of Plackett-Luce models.
}
\references{
Mollica, C. and Tardella, L. (2017). Bayesian Plackett-Luce mixture models for partially ranked data. \emph{Psychometrika}, \bold{82}(2), pages 442--458, ISSN: 0033-3123, <doi:10.1007/s11336-016-9530-0>.

Mollica, C. and Tardella, L. (2014). Epitope profiling via mixture modeling for ranked data. \emph{Statistics in Medicine}, \bold{33}(21), pages 3738--3758, ISSN: 0277-6715, <doi:10.1002/sim.6224>.
}
\author{
Cristina Mollica and Luca Tardella
}
