% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quickintsample}
\alias{quickintsample}
\title{Weighted sampling without replacement from a finite urn}
\usage{
quickintsample(n, size, prob)
}
\arguments{
\item{n}{Number of distinct integer-labeled balls in the urn.}

\item{size}{Number of balls sampled without replacement.}

\item{prob}{Numeric vector of length \eqn{n} probability masses assigned to each ball.}
}
\value{
vector
}
\description{
Weighted sampling without replacement from a finite urn
}
