% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimYpsilon}
\alias{SimYpsilon}
\title{Gibbs sampling of the quantitative latent variables}
\usage{
SimYpsilon(rate, n_rank)
}
\arguments{
\item{rate}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} matrix of rate parameters.}

\item{n_rank}{Numeric vector of length \eqn{N} with the number of items ranked by each sample unit.}
}
\value{
Numeric \eqn{N}\eqn{\times}{x}\eqn{K} matrix of posterior samples of the quantitative latent variables.
}
\description{
The function \code{CompRateYpartial} simulates from the Exponential full-conditionals of the quantitative latent variables for the Gibbs sampling of a Bayesian mixture of Plackett-Luce models.
}
