% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tau}
\alias{tau}
\title{Compute paired comparison matrix for a partial ordering dataset}
\usage{
tau(pi_inv)
}
\arguments{
\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}
}
\value{
The \eqn{K}\eqn{\times}{x}\eqn{K} paired comparison matrix: number of times that item \eqn{i} is preferred to item \eqn{i'}.
}
\description{
Compute paired comparison matrix for a partial ordering dataset
}
