% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{umat}
\alias{umat}
\title{Binary matrix detailing the items ranked by each sample unit}
\usage{
umat(pi_inv)
}
\arguments{
\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}
}
\value{
Binary \eqn{N}\eqn{\times}{x}\eqn{K} matrix indicating whether sample unit \eqn{s} ranked item \eqn{i}.
}
\description{
The function \code{umat} returns a binary matrix whose elements indicate which items has been ranked by each sample unit.
}
