% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMIXfunctions.R
\name{unit_to_freq}
\alias{unit_to_freq}
\title{Frequency distribution from the individual rankings/orderings}
\usage{
unit_to_freq(data)
}
\arguments{
\item{data}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of observed individual sequences.}
}
\value{
Numeric matrix of the distinct observed sequences with the corresponding frequencies indicated in the last \eqn{(K+1)}-th column.
}
\description{
Construct the frequency distribution of the distinct observed sequences from the dataset of individual rankings/orderings.
}
\examples{

## Frequency distribution for the APA top-ordering dataset
data(d_apa)
unit_to_freq(data=d_apa)
}
\author{
Cristina Mollica and Luca Tardella
}
