% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpPhetpartial}
\alias{UpPhetpartial}
\title{M-step for the support parameters of a Bayesian mixture of Plackett-Luce models}
\usage{
UpPhetpartial(p, ref_order, pi_inv, u_bin, z_hat, shape0, rate0, n_rank)
}
\arguments{
\item{p}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific support parameters.}

\item{ref_order}{Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of component-specific reference orders.}

\item{pi_inv}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of partial orderings.}

\item{u_bin}{Binary \eqn{N}\eqn{\times}{x}\eqn{K} matrix indicating whether sample unit \eqn{s} ranked item \eqn{i}.}

\item{z_hat}{Numeric \eqn{N}\eqn{\times}{x}\eqn{G} matrix of posterior component membership probabilities.}

\item{shape0}{Numeric \eqn{G}\eqn{\times}{x}{K} matrix of shape hyperparameters.}

\item{rate0}{Numeric vector of \eqn{G} rate hyperparameters.}

\item{n_rank}{Numeric vector of length \eqn{N} with the number of items ranked by each sample unit.}
}
\value{
Numeric \eqn{G}\eqn{\times}{x}\eqn{K} matrix of estimated component-specific support parameters.
}
\description{
The function \code{UpPhetpartial} updates the support parameter estimates in the EM algorithm for MAP estimation of a Bayesian mixture of Plackett-Luce models.
}
