% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpWhet}
\alias{UpWhet}
\title{M-step for the weights of a Bayesian mixture of Plackett-Luce models}
\usage{
UpWhet(z_hat, alpha0)
}
\arguments{
\item{z_hat}{Numeric \eqn{N}\eqn{\times}{x}\eqn{G} matrix of posterior component membership probabilities.}

\item{alpha0}{Numeric vector of \eqn{G} Dirichlet hyperparameters.}
}
\value{
Numeric vector of the \eqn{G} estimated mixture weights.
}
\description{
The function \code{UpWhet} updates the mixture weight estimates in the EM algorithm for MAP estimation of a Bayesian mixture of Plackett-Luce models
}
