% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLMIXfunctions.R
\name{unit_to_freq}
\alias{unit_to_freq}
\title{Frequency distribution from the individual rankings/orderings.}
\usage{
unit_to_freq(data)
}
\arguments{
\item{data}{Numeric \eqn{N}\eqn{\times}{x}\eqn{K} data matrix of observed individual sequences.}
}
\value{
Numeric matrix of the observed sequences with the corresponding frequencies indicated in the last \eqn{(K+1)}-th column.
}
\description{
Construct the frequency distribution of the observed sequences from the dataset of individual rankings/orderings.
}
\examples{

library(PLMIX)
data(d_german)
unit_to_freq(data=d_german)

}
\author{
Cristina Mollica and Luca Tardella
}

